services/ai-core/pyproject.toml
[project]
name = "ai-search-assistant-core"
version = "0.1.0"
description = "Python AI core service for ai-search-assistant (RAG + orchestration)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "httpx>=0.27.0",
  "qdrant-client>=1.14.0",
  "fastembed>=0.4.0",
  "opentelemetry-api>=1.27.0,<2",
  "opentelemetry-sdk>=1.27.0,<2",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0,<2",
  "opentelemetry-instrumentation-fastapi>=0.48b0,<1",
  "opentelemetry-instrumentation-httpx>=0.48b0,<1",
  "pypdf>=5.0.0",
  "python-docx>=1.1.0",
  "python-pptx>=1.0.0",
  "striprtf>=0.0.26",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
  "aiosqlite>=0.20.0",
]
sqlite = ["aiosqlite>=0.20.0"]
mysql = ["aiomysql>=0.2.0"]
all-databases = ["aiosqlite>=0.20.0", "aiomysql>=0.2.0"]

[project.scripts]
ai-search-assistant-ingest = "ai_search_assistant.ingestion.ingest_cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
